requirements.txt
l5kit==1.5.0
numpy==1.19.5
Pillow==8.4.0
tensorflow==2.7.0
